Extract tournament handlers to separate module
